Extract ButtonSlider styled component into its own file.
